refactor: 로긍니 페이지 컴포넌트 분리
